Add Stock Firmware from VM Evo V
Added stock Evo V 4g firmware for BT from last OTA update along with changes to the Kernel defconfig. Am seeing a noticeable difference.

Not sure about needing “bcm4330.hcd” but have copied it over as it is included in “msm8660-common” so will do some testing with and without.